dev-requirements.txt
nose
PyHamcrest
